+++ src/main/java/in/koreatech/koin/admin/operators/dto/response/AdminTokenRefreshResponse.java
@@ -1,4 +1,4 @@
-package in.koreatech.koin.admin.user.dto;
+package in.koreatech.koin.admin.operators.dto.response;
 
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
 
@@ -6,7 +6,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
-public record AdminLoginResponse (
+public record AdminTokenRefreshResponse (
     @Schema(
         description = "Jwt accessToken",
         example = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c",
@@ -15,12 +15,12 @@ public record AdminLoginResponse (
     @JsonProperty("token")
     String accessToken,
 
-    @Schema(description = "Random UUID refresh token", example = "RANDOM-KEY-VALUE", requiredMode = REQUIRED)
+    @Schema(description = "Random UUID refreshToken", example = "RANDOM-KEY-VALUE", requiredMode = REQUIRED)
     @JsonProperty("refresh_token")
     String refreshToken
 ) {
 
-    public static AdminLoginResponse of(String token, String refreshToken) {
-        return new AdminLoginResponse(token, refreshToken);
+    public static AdminTokenRefreshResponse of(String accessToken, String refreshToken) {
+        return new AdminTokenRefreshResponse(accessToken, refreshToken);
     }
 }
