+++ src/main/java/in/koreatech/koin/admin/user/controller/AdminUserController.java
@@ -17,17 +17,17 @@ import org.springframework.web.bind.annotation.RestController;
 import in.koreatech.koin.global.auth.Auth;
 import in.koreatech.koin.admin.abtest.useragent.UserAgent;
 import in.koreatech.koin.admin.abtest.useragent.UserAgentInfo;
-import in.koreatech.koin.admin.user.dto.AdminLoginRequest;
-import in.koreatech.koin.admin.user.dto.AdminLoginResponse;
-import in.koreatech.koin.admin.user.dto.AdminPasswordChangeRequest;
-import in.koreatech.koin.admin.user.dto.AdminPermissionUpdateRequest;
-import in.koreatech.koin.admin.user.dto.AdminResponse;
-import in.koreatech.koin.admin.user.dto.AdminTokenRefreshRequest;
-import in.koreatech.koin.admin.user.dto.AdminTokenRefreshResponse;
-import in.koreatech.koin.admin.user.dto.AdminUpdateRequest;
-import in.koreatech.koin.admin.user.dto.AdminsCondition;
-import in.koreatech.koin.admin.user.dto.AdminsResponse;
-import in.koreatech.koin.admin.user.dto.CreateAdminRequest;
+import in.koreatech.koin.admin.operators.dto.request.AdminLoginRequest;
+import in.koreatech.koin.admin.operators.dto.response.AdminLoginResponse;
+import in.koreatech.koin.admin.operators.dto.request.AdminPasswordChangeRequest;
+import in.koreatech.koin.admin.operators.dto.request.AdminPermissionUpdateRequest;
+import in.koreatech.koin.admin.operators.dto.response.AdminResponse;
+import in.koreatech.koin.admin.operators.dto.request.AdminTokenRefreshRequest;
+import in.koreatech.koin.admin.operators.dto.response.AdminTokenRefreshResponse;
+import in.koreatech.koin.admin.operators.dto.request.AdminUpdateRequest;
+import in.koreatech.koin.admin.operators.dto.request.AdminsCondition;
+import in.koreatech.koin.admin.operators.dto.response.AdminsResponse;
+import in.koreatech.koin.admin.operators.dto.response.CreateAdminRequest;
 import in.koreatech.koin.admin.user.enums.TeamType;
 import in.koreatech.koin.admin.user.enums.TrackType;
 import in.koreatech.koin.admin.user.service.AdminUserService;
