+++ src/main/java/in/koreatech/koin/admin/user/repository/AdminRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.Repository;
 import org.springframework.data.repository.query.Param;
 
-import in.koreatech.koin.admin.user.dto.AdminsCondition;
+import in.koreatech.koin.admin.operators.dto.request.AdminsCondition;
 import in.koreatech.koin.admin.user.exception.AdminNotFoundException;
 import in.koreatech.koin.admin.user.model.Admin;
