+++ src/main/java/in/koreatech/koin/admin/user/service/AdminUserService.java
@@ -14,17 +14,17 @@ import in.koreatech.koin.common.model.Criteria;
 import in.koreatech.koin.admin.abtest.useragent.UserAgentInfo;
 import in.koreatech.koin.admin.owner.repository.AdminOwnerRepository;
 import in.koreatech.koin.admin.student.repository.AdminStudentRepository;
-import in.koreatech.koin.admin.user.dto.AdminLoginRequest;
-import in.koreatech.koin.admin.user.dto.AdminLoginResponse;
-import in.koreatech.koin.admin.user.dto.AdminPasswordChangeRequest;
-import in.koreatech.koin.admin.user.dto.AdminPermissionUpdateRequest;
-import in.koreatech.koin.admin.user.dto.AdminResponse;
-import in.koreatech.koin.admin.user.dto.AdminTokenRefreshRequest;
-import in.koreatech.koin.admin.user.dto.AdminTokenRefreshResponse;
-import in.koreatech.koin.admin.user.dto.AdminUpdateRequest;
-import in.koreatech.koin.admin.user.dto.AdminsCondition;
-import in.koreatech.koin.admin.user.dto.AdminsResponse;
-import in.koreatech.koin.admin.user.dto.CreateAdminRequest;
+import in.koreatech.koin.admin.operators.dto.request.AdminLoginRequest;
+import in.koreatech.koin.admin.operators.dto.response.AdminLoginResponse;
+import in.koreatech.koin.admin.operators.dto.request.AdminPasswordChangeRequest;
+import in.koreatech.koin.admin.operators.dto.request.AdminPermissionUpdateRequest;
+import in.koreatech.koin.admin.operators.dto.response.AdminResponse;
+import in.koreatech.koin.admin.operators.dto.request.AdminTokenRefreshRequest;
+import in.koreatech.koin.admin.operators.dto.response.AdminTokenRefreshResponse;
+import in.koreatech.koin.admin.operators.dto.request.AdminUpdateRequest;
+import in.koreatech.koin.admin.operators.dto.request.AdminsCondition;
+import in.koreatech.koin.admin.operators.dto.response.AdminsResponse;
+import in.koreatech.koin.admin.operators.dto.response.CreateAdminRequest;
 import in.koreatech.koin.admin.user.model.Admin;
 import in.koreatech.koin.admin.user.repository.AdminRepository;
 import in.koreatech.koin.admin.user.repository.AdminUserRepository;
