+++ src/main/java/in/koreatech/koin/admin/user/validation/AdminUserValidation.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Component;
 import in.koreatech.koin.global.auth.exception.AuthorizationException;
 import in.koreatech.koin.global.exception.CustomException;
 import in.koreatech.koin.global.code.ApiResponseCode;
-import in.koreatech.koin.admin.user.dto.AdminLoginRequest;
+import in.koreatech.koin.admin.operators.dto.request.AdminLoginRequest;
 import in.koreatech.koin.admin.user.repository.AdminRepository;
 import in.koreatech.koin.admin.user.repository.AdminUserRepository;
 import in.koreatech.koin.domain.user.model.User;
